Replace texture atlases with custom resources
Storing the mapping from an input to a texture in a resource simplifies
the code and makes it more flexible; it will be easier to respond to
future changes in Godot's mappings, without having to adjust atlas
offsets.

The new resources are compatible with Godot 4.0's renumbering of joypad
and mouse buttons. Closes #1.