Add function to refresh input prompts
Calling the refresh() function of a specific prompt node forces it to
update its events and icons. Calling the refresh() function of the
PromptManager singleton refreshes all input prompts.

refresh() must be called whenever the events associated with an action
in the InputMap changes, because Godot does not provide an ability to
detect changes in the InputMap.

Closes #8.